requirements.txt
flask==2.2.3
tensorflow-cpu==2.10.0
numpy==1.23.5
Pillow==9.4.0
requests==2.28.2
gunicorn==20.1.0
python-dotenv==1.0.0
scikit-learn==1.2.2
werkzeug==2.2.3
boto3==1.26.0  # For AWS S3
SQLAlchemy==2.0.0  # For PostgreSQL
psycopg2-binary==2.9.5  # PostgreSQL adapter